Add 'unknown' contacts concept/feature (and filters for it).
* add 'is_unknown' field to contact/person DB entries:

Add is_unknown support to contacts-service2. This patch introduces the
the basic infrastructure and the 'db_plugin_contact' support to
insert/select the aforementioned field of contacts/persons. The field
is designed to flag contacts which were not *actively* added into the
contacts database by an user, but rather by an application by which
the user was contacted, from a contact (number, email, etc) not yet
present on the contacts pristine listing.

* filter out all the unknown contacts:

Change all the queries used for listing contacts to filter out the
unknown contacts. List just those included by the user. This way, the
patch is totally no intrusive to anyone not willing to query/deal with
unknown contacts.

* introduce a contact view including unknown contacts

Add a new database view to select all contacts, including the unknown
ones.

* introduce a person view including unknown contacts

The same function as above, but on person level, not contacts.

Contacts marked with 'is_unknown' not zero get an integer on the
field, representing the time of its creation. A service (or similar
piece of code) can be run from time to time, dumping unknown contacts
older than a specified amount of time, if need be.

Change-Id: Ia26c301ff1ffe6b45a6743250b0f98a8017c7142
Signed-off-by: Gustavo Lima Chaves <gustavo.lima.chaves@intel.com>